refactor: move iluvatar resource resolution into backend util
